build: add clang tidy to enforce modern c++ usage (#591)
# Pull Request

## Description

To ensure that the codebase follows good principles like modern usage of
cpp, I added clang tidy configuration file with an automatic check
inside CI.

## Related Issues (Put "None" if there are no related issues)

close #553 

## Type of Change
Please delete options that are not relevant.

- Build/CI configuration change

## Changes Made
List the main changes in this PR:
- Added clang tidy config file
- Add a gate in CI to check the code conformity to clang tidy

## Testing
Describe the tests you ran to verify your changes. Please delete options
that are not relevant.
- Unit tests pass (`xmake test`)

### Test Environment
- OS: macOS
- Compiler: Clang

## Screenshots/Videos (Put "None" if there are no related issues)

None

## Documentation
Please delete options that are not relevant.

- No documentation changes are required

## Checklist (Don't delete any options)
- [x] My code follows the style guidelines of this project
- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] My changes generate no new warnings
- [x] I have added tests that prove my fix is effective or that my
feature works
- [x] New and existing unit tests pass locally with my changes
- [x] Any dependent changes have been merged and published

## Breaking Changes (Put "None" if there are no related issues)

None

## Additional Notes (Put "None" if there are no related issues)

None

+++ src/plugin/object/src/utils/SphereGenerator.cpp
@@ -19,7 +19,7 @@ Component::Mesh GenerateSphereMesh(float radius, uint32_t segments, uint32_t rin
     mesh.ReserveVertices(vertexCount);
     mesh.ReserveNormals(vertexCount);
     mesh.ReserveTexCoords(vertexCount);
-    mesh.ReserveIndices(rings * segments * 6u);
+    mesh.ReserveIndices(static_cast<size_t>(rings * segments) * 6u);
 
     for (uint32_t ring = 0u; ring <= rings; ++ring)
     {
